+++ src/strip/JoinedStrip.cpp
@@ -351,3 +351,18 @@ bool JoinedStrip::paintRandomPos(int16_t length, CRGB color, bool add) {
     uint16_t pos = random16(last() - length);
     return paint(pos, pos + length, color, add);
 }
+
+CRGB JoinedStrip::getIndex(int16_t index) {
+    if (isInStrip1(index)) {
+        return strip1->getIndex(toStrip1(index));
+    } else if (isInGap(index)) {
+        return gapStrip->getIndex(toGap(index));
+    } else if (isInStrip2(index)) {
+        return strip2->getIndex(toStrip2(index));
+    }
+    return CRGB::Black;
+}
+
+CRGB JoinedStrip::getPosition(double position) {
+    return getIndex(fromNormalizedPosition(position));
+}